% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{BinToDec}
\alias{BinToDec}
\title{Convert a binary number to integer}
\usage{
BinToDec(x)
}
\arguments{
\item{x}{the binary number}
}
\value{
an integer
}
\description{
Convert a binary number to integer
}
