% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_recovery.R
\name{calcFrazier}
\alias{calcFrazier}
\title{Calculate recovery metrics from a time series with known disturbance date. The calcFrazier function derives the RRI, R80P and YrYr recovery indicators, defined by Frazier et al. (2018). The indicators are originally developped for annual long-term time series of optical vegetation indices (the indicators are shown in the figures below). Yet, in order to be able to derive the indicators as well for dense and/or short time series, a modified version is suggested. Here, the user can define the time period before, during and after the disturbance that is used to derive the indicators. To reduce the interference of the seasonal pattern of dense time series, the chosen time period should cover blocks of n years. Moreover, given the potentially high noise levels of dense time series, the mean value instead of the maximum value was used in the formulas. (Frazier, R. J., Coops, N. C., Wulder, M. A., Hermosilla, T., & White, J. C. (2018). Analyzing spatial and temporal variability in short-term rates of post-fire vegetation return from Landsat time series. Remote Sensing of Environment, 205, 32-45.)}
\usage{
calcFrazier(tsio, tdist, obspyr, shortDenseTS, nPre, nDist, nPostMin,
  nPostMax)
}
\arguments{
\item{tsio}{vector of observations (time series with a fixed observation frequency)}

\item{tdist}{observation number of disturbance, indicating the timing of the disturbance}

\item{obspyr}{number of observations per year}

\item{shortDenseTS}{TRUE or FALSE. In case TRUE, the metrics are adjusted to be compatible with short, dense time series}

\item{nPre}{If shortDenseTS is TRUE, number of years prior to the disturbance used to calculate the pre-disturbance value}

\item{nDist}{If shortDenseTS is TRUE, number of months used to quantify the time series value during the disturbance}

\item{nPostMin}{If shortDenseTS is TRUE,  the post-disturbance condition is quantified starting from nPostMin years after the disturbance}

\item{nPostMax}{If shortDenseTS is TRUE, max number of years after the disturbance used to quantify the post-disturbance condition}
}
\value{
a list containing the RRI recovery indicator, R80p recovery indicator and YrYr recovery indicator
}
\description{
Calculate recovery metrics from a time series with known disturbance date. The calcFrazier function derives the RRI, R80P and YrYr recovery indicators, defined by Frazier et al. (2018). The indicators are originally developped for annual long-term time series of optical vegetation indices (the indicators are shown in the figures below). Yet, in order to be able to derive the indicators as well for dense and/or short time series, a modified version is suggested. Here, the user can define the time period before, during and after the disturbance that is used to derive the indicators. To reduce the interference of the seasonal pattern of dense time series, the chosen time period should cover blocks of n years. Moreover, given the potentially high noise levels of dense time series, the mean value instead of the maximum value was used in the formulas. (Frazier, R. J., Coops, N. C., Wulder, M. A., Hermosilla, T., & White, J. C. (2018). Analyzing spatial and temporal variability in short-term rates of post-fire vegetation return from Landsat time series. Remote Sensing of Environment, 205, 32-45.)
}
