% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{calcNBR}
\alias{calcNBR}
\title{Calculate Normalized Burn Ratio}
\usage{
calcNBR(NIR, SWIR)
}
\arguments{
\item{NIR}{Reflectance NIR band (band 4 for landsat 4-7 and band 5 for landsat 8)}

\item{SWIR}{Reflectance SWIR band (band 7 of landsat)}
}
\value{

}
\description{
Calculate Normalized Burn Ratio
}
