% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_recovery.R
\name{calcRecoveryTS}
\alias{calcRecoveryTS}
\title{Calculate recovery for a single time series}
\usage{
calcRecoveryTS(
  tsi,
  maxBreak,
  obspyr,
  inp = "segmented",
  shortDenseTS = TRUE,
  nPre = 2,
  nDist = 12,
  nPostMin = 4,
  nPostMax = 6,
  h = 0.15,
  timeThres,
  slpThres
)
}
\arguments{
\item{tsi}{vector: the first n values contain the timing of the disturbances and the next n values the observations for which the recovery indicators should be computed}

\item{maxBreak}{(only for recovery indicators derived from segmented time series): if maxbreak is true, the maximum break in the segmented series is used as disturbance date to calculate the recovery indicators. If maxbreak is false, the break closest to the provided disturbance timing is used to calculate recovery.}

\item{obspyr}{the number of observations per year}

\item{inp}{the preprocessing applied to the time series before computing the recovery indicators: segmented (for segmentations), smooth (time series smoothing using loess), or raw (no preprocessing)}

\item{shortDenseTS}{if true, the recovery indicators are user designed, i.e. they can be adjusted to deal with short and/or dense time series}

\item{nPre}{number of years prior to the disturbance that are used to derive the pre-disturbance condition}

\item{nDist}{number of observations used to derive the disturbance state}

\item{nPostMin}{start of the post-disturbance period: number of years after the disturbance}

\item{nPostMax}{end of the post-disturbance period: number of years after the disturbance}

\item{h}{h parameter of the breakpoints function in the strucchange package}

\item{timeThres}{only relevant for segmentation: threshold on the duration between the disturbance date and date of the detected break [years]}

\item{slpThres}{only relevant for segmentation:threshold on the pre-disturbance slope}
}
\value{
the RRI, R80P, YrYr and the slope of the pos-disturbance segment
}
\description{
Calculate recovery for a single time series
}
