% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_recovery.R
\name{calcSegRec}
\alias{calcSegRec}
\title{Post-disturbance slope and recovery metrics derived from BFAST0n trend segments. The calcBFASTrec function derives a set of recovery indicators after fitting a segmented trend in the time series. Using the breakpoints function of the strucchange package, a segmented trend is fitted (hereafter called BFAST0n trend segments). The detected break showing the largest change (in absolute values) is assumed to represent the disturbance. Using the segmented trend and detected disturbance date, the RRI, R80p, YrYr and the slope of the post-disturbance trend segment are derived as recovery indicators.}
\usage{
calcSegRec(
  tsio,
  tdist,
  maxBreak,
  obspyr,
  h,
  shortDenseTS,
  nPre,
  nDist,
  nPostMin,
  nPostMax,
  timeThres,
  seas = F
)
}
\arguments{
\item{tsio}{vector of observations (time series)}

\item{tdist}{the timing of the disturbance [observation number]}

\item{maxBreak}{only for recovery indicators derived from piecewise regression: if maxbreak is true, the maximum break in the segmented series is used as disturbance date to calculate the recovery indicators. If maxbreak is false, the break closest to the provided disturbance timing is used to calculate recovery.}

\item{obspyr}{number of observations in one year}

\item{h}{This parameter defines the minimal segment size either given as fraction relative to the sample size or as an integer giving the minimal number of observations in each segment.}

\item{shortDenseTS}{TRUE or FALSE. In case FALSE, the metrics follow closely the definitions given by Frazier et al}

\item{nPre}{number of years prior to the disturbance used to calculate the pre-disturbance value}

\item{nDist}{number of years used to quantify the time series value during the disturbance}

\item{nPostMin}{min number of years after the disturbance used to quantify the recovery}

\item{nPostMax}{max number of years after the disturbance used to quantify the recovery}

\item{timeThres}{threshold on the duration between the disturbance date and date of the detected break [years]}

\item{seas}{should a seasonal comonent be used in the piecewise regression?}
}
\value{
a list containing  the RRI, R80p, YrYr recovery indicator derived from the BFAST0n trend segments and slope of the trend segment after the disturbance (sl).
}
\description{
Post-disturbance slope and recovery metrics derived from BFAST0n trend segments. The calcBFASTrec function derives a set of recovery indicators after fitting a segmented trend in the time series. Using the breakpoints function of the strucchange package, a segmented trend is fitted (hereafter called BFAST0n trend segments). The detected break showing the largest change (in absolute values) is assumed to represent the disturbance. Using the segmented trend and detected disturbance date, the RRI, R80p, YrYr and the slope of the post-disturbance trend segment are derived as recovery indicators.
}
