% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_generator.R
\name{exponential}
\alias{exponential}
\title{Exponential decay}
\usage{
exponential(t, offset = 0, pert = 0, tpert = 0, thalf = 1, noise = 0)
}
\arguments{
\item{t}{Times to simulate}

\item{offset}{Offset}

\item{pert}{Perturbation intensity (signed)}

\item{tpert}{Perturbation timing}

\item{thalf}{Perturbation half-life}

\item{noise}{Strength of the additive white noise (standard deviation)}
}
\value{
The time series
}
\description{
Exponential decay
}
