% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{getGrid}
\alias{getGrid}
\title{Extract the extent of each grid tile that covers an area of interest}
\usage{
getGrid(cubefolder, ext)
}
\arguments{
\item{ext}{extent of the area of interest}

\item{l2folder}{directory of the level2 data cube}
}
\value{
list of extents
}
\description{
Extract the extent of each grid tile that covers an area of interest
}
