% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{getQA}
\alias{getQA}
\title{Generate the QA value associated with specific quality settings}
\usage{
getQA(
  valid = 0,
  cloud = 0,
  shadow = 0,
  snow = 0,
  water = 0,
  aero = 0,
  subzero = 0,
  sat = 0,
  sunZen = 0,
  illum = 0,
  slope = 0,
  wvp = 0
)
}
\arguments{
\item{valid}{quality parameter - valid data, values can be 0 (valid) or 1 (no data)}

\item{cloud}{quality parameter - cloud state, values can be 0 (clear), 1 (less confident cloud, i.e. buffered cloud 300m), 2 (confident, opaque cloud), or 3 (cirrus)}

\item{shadow}{quality parameter - cloud shadow flag, values can be 0 (no shadow), or 1 (shadow)}

\item{snow}{quality parameter - snow flag, values can be 0 (no snow), or 1 (snow)}

\item{water}{quality parameter - water flag, values can be 0 (no water), or 1 (water)}

\item{aero}{quality parameter - aerosol state, values can be 0 (estimated, best quality), 1 (interpolated, mid quality), 2 (high, aerosol optical depth > 0.6, use with caution)}

\item{subzero}{quality parameter - subzero flag, values can be 0 (no subzero values), or 1 (subzero values)}

\item{sat}{quality parameter - saturation flag, values can be 0 (no saturation), or 1 (saturation)}

\item{sunZen}{quality parameter - high sun zenith flag, values can be 0 (no high sun zenith), or 1 (sun elevation < 15 degrees, use with caution)}

\item{illum}{quality parameter - illumination state, values can be 0 (good, incidence angle <55°, best quality for topographic correction), 1 (medium, incidence angle 55°-80°, low quality for top. correction), 2 (poor, incidence angle > 80°, low quality for top. correction), or 3 (shadow, incidence angle > 90°, no topographic correction applied)}

\item{slope}{quality parameter - slope flag, values can be 0 (cosine correction applied) or 1 (enchanced C-correction applied)}

\item{wvp}{quality parameter - water vapor flag, values can be 0 (measured, best quality, only for Sentinel-2) or 1 (fill, scene average, only for Sentinel-2)}
}
\value{
QA value for the given parameter value setting
}
\description{
Generate the QA value associated with specific quality settings
}
