% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{maskAOI}
\alias{maskAOI}
\title{Set terra raster values outside area of interest to 0}
\usage{
maskAOI(fmask, ext)
}
\arguments{
\item{fmask}{a terra raster layer}

\item{ext}{extent of area of interest (vector with xmin, xmax, ymin, ymax), expressed as longitude and latitude}
}
\value{
terra raster layer
}
\description{
Set terra raster values outside area of interest to 0
}
