% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{maskL47}
\alias{maskL47}
\title{Mask low quality observations for Landsat 4, 5, or 7 time series.}
\usage{
maskL47(x)
}
\arguments{
\item{x}{a vector, where the first n observations represent Landsat observations,
the next n observations represent the associated SR_CLOUD_QA values,
the next n observations represent the QA_PIXEL values,
and the final n observations represent the QA_RADSAT values}
}
\value{
vector of Landsat observations for which low quality observations are masked
}
\description{
Mask low quality observations for Landsat 4, 5, or 7 time series.
}
