% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{maskL47Stack}
\alias{maskL47Stack}
\title{Mask low quality observations for a stack of Landsat 4, 5, or 7 images.}
\usage{
maskL47Stack(x)
}
\arguments{
\item{x}{a raster stack, where the first image is a mask (only pixels with a value equal to 1 are processed),
the subsequent n images represent Landsat observations,
the next n images represent the QA_PIXEL values,
and the final n images represent the QA_RADSAT values}
}
\value{
stack of images for which low quality observations are masked
}
\description{
Mask low quality observations for a stack of Landsat 4, 5, or 7 images.
}
