% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{maskL8}
\alias{maskL8}
\title{Mask low quality observations for Landsat 8 time series.}
\usage{
maskL8(x)
}
\arguments{
\item{x}{a vector, where the first n observations represent Landsat observations,
the next n observations represent the QA_PIXEL values,
and the final n observations represent the QA_RADSAT values}
}
\value{
vector of Landsat observations for which low quality observations are masked
}
\description{
Mask low quality observations for Landsat 8 time series.
}
