% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{mmean}
\alias{mmean}
\title{Helper function for the toRegularTSStack function}
\usage{
mmean(x)
}
\arguments{
\item{x}{vector of observations}
}
\value{
the mean value of the vector
}
\description{
Helper function for the toRegularTSStack function
}
