% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{prepareNBRstack}
\alias{prepareNBRstack}
\title{Title}
\usage{
prepareNBRstack(
  img_list,
  qai_list,
  fmask,
  lsens,
  ldts,
  tempFun,
  tempRes,
  startdt,
  enddt
)
}
\arguments{
\item{endtime}{}
}
\value{

}
\description{
Title
}
