% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{r80p}
\alias{r80p}
\title{R80p recovery function}
\usage{
r80p(ts, ys, r = 0.8, ts_pre = c(-1, -2), ts_post = c(4, 5))
}
\arguments{
\item{ts}{Vector containing the times (same size as ys). Perturbation assumed to happen at 0}

\item{ys}{Vector containing the values (same size as ts)}

\item{r}{Ratio. Default set to 0.8}

\item{ts_pre}{Sampling times for estimating Vpre. Default set to -1}

\item{ts_post}{Sampling times for estimating Vpost. Default set to c(4, 5)}
}
\value{
The R80p indicator (R_r_p if r != 0.8 is provided)
}
\description{
Ratio of Eighty Percent (R80P). The ratios can be customized.
}
\examples{
# Generate an example time series
ts <- seq(-2, 10, by = 0.1) # as a vector of times
ys <- exponential(ts, pert = -2, offset = 1, thalf = 0.25) # plus a vector of values
r80p(ts, ys)
}
\references{
Frazier, R. J., Coops, N. C., Wulder, M. A., Hermosilla, T., & White, J. C. (2018).
Analyzing spatial and temporal variability in short-term rates of post-fire vegetation return
from Landsat time series. Remote Sensing of Environment, 205, 32–45.
\url{https://doi.org/10.1016/j.rse.2017.11.007}
}
