% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_generator.R
\name{realistic}
\alias{realistic}
\title{Realistic time series simulation}
\usage{
realistic(t, offset = 0, pert = 0, tpert = 0, thalf = 1, noise = 0)
}
\arguments{
\item{t}{Times to simulate}

\item{offset}{Offset}

\item{pert}{Perturbation intensity (signed)}

\item{tpert}{Perturbation timing}

\item{thalf}{Perturbation half-life}

\item{noise}{Strength of the stochastic term in the differential equation (standard deviation of the integrated time series)}
}
\value{
The time series
}
\description{
Simulates a return to equilibrium after a perturbation under the influence of a stochastic differential equation where:
}
\details{
1. The deterministic dynamics are given by an exponential decay with the given half life
2. The stochastic dynamics have the given infinitesimal standard deviation
3. Gaussian noise is added to the final time series to simulate measurement errors
}
