% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{rollback_quart}
\alias{rollback_quart}
\title{Rollback dates to first day of quarterly series}
\usage{
rollback_quart(dts)
}
\arguments{
\item{dts}{dates to be converted (date object)}
}
\value{
date object
}
\description{
Rollback dates to first day of quarterly series
}
