% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{rri}
\alias{rri}
\title{RRI recovery function}
\usage{
rri(ts, ys, tpert = 0, ts_pre = -1, ts_post = c(4, 5))
}
\arguments{
\item{ts}{Vector containing the times (same size as ys). Perturbation assumed to happen at 0}

\item{ys}{Vector containing the values (same size as ts)}

\item{tpert}{Time of the perturbation. Default set to 0 yr}

\item{ts_pre}{Sampling times for estimating Vpre. Default set to -1}

\item{ts_post}{Sampling times for estimating Vpost. Default set to c(4, 5)}
}
\value{
The RRI indicator
}
\description{
Relative Recovery Indicator
}
\examples{
# Generate an example time series
ts <- seq(-2, 10, by = 0.1) # as a vector of times
ys <- exponential(ts, pert = -2, offset = 1, thalf = 0.25) # plus a vector of values
rri(ts, ys)
}
\references{
Frazier, R. J., Coops, N. C., Wulder, M. A., Hermosilla, T., & White, J. C. (2018).
Analyzing spatial and temporal variability in short-term rates of post-fire vegetation return
from Landsat time series. Remote Sensing of Environment, 205, 32–45.
\url{https://doi.org/10.1016/j.rse.2017.11.007}
}
