% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{setPeriod}
\alias{setPeriod}
\title{Fix the time span of a multi-temporal raster stack to a user defined time window}
\usage{
setPeriod(br, starttime, endtime, tempRes, dtsbr)
}
\arguments{
\item{br}{raster stack}

\item{starttime}{start date of the desired time span (given as a numeric vector containing the year, month and day)}

\item{endtime}{end date of the desired time span (given as a numeric vector containing the year, month and day)}

\item{tempRes}{temporal resolution of the raster stack}

\item{dtsbr}{vector of dates (Date object) associated with the raster stack}
}
\value{
raster stack with adjusted time span
}
\description{
Fix the time span of a multi-temporal raster stack to a user defined time window
}
