% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{toAnnualTS}
\alias{toAnnualTS}
\title{Convert time series to annual frequency: The toAnnualTS function converts a time series with n observations per year to an annual time series (one observation per year). The main concept is to select observations per year closest to a given day of year that have no missing value (NA). Here, the day of year for which the seasonality is maximum is being used.}
\usage{
toAnnualTS(tsseas, tsi, obspyr)
}
\arguments{
\item{tsseas}{vector of observations (time series) representing the seasonal component of the time series to be converted}

\item{tsi}{vector of observations (time series) that needs to be converted to an annual time series}

\item{obspyr}{number of observations per year of the time series to be converted}
}
\value{
vector of observations (time series) with annual observation frequency
}
\description{
Convert time series to annual frequency: The toAnnualTS function converts a time series with n observations per year to an annual time series (one observation per year). The main concept is to select observations per year closest to a given day of year that have no missing value (NA). Here, the day of year for which the seasonality is maximum is being used.
}
