% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{TScompress}
\alias{TScompress}
\title{The TScompress function compresses a time series vector. The purpose is to avoid storing many NA values.}
\usage{
TScompress(ts)
}
\arguments{
\item{ts}{a vector that will be compressed}
}
\value{
a vector containing, in order, the length of the input vector, the number of observations without NA, the observations without NA, and the values of the observations that are no NA
}
\description{
The TScompress function compresses a time series vector. The purpose is to avoid storing many NA values.
}
