% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{TSdecompress}
\alias{TSdecompress}
\title{The TSdecompress recovers a vector that has been compressed by the TScompress function in its original format.}
\usage{
TSdecompress(ts)
}
\arguments{
\item{ts}{a vector compressed be the TScompress function}
}
\value{
the vector restored in its original format
}
\description{
The TSdecompress recovers a vector that has been compressed by the TScompress function in its original format.
}
