% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{yryr}
\alias{yryr}
\title{YrYr recovery function}
\usage{
yryr(ts, ys, tpert = 0, deltat = 5)
}
\arguments{
\item{ts}{Vector containing the times (same size as ys)}

\item{ys}{Vector containing the values (same size as ts)}

\item{tpert}{Time of the perturbation. Default set to 0 yr}

\item{deltat}{Reference recovery time (in years). Default set to 5 yr}
}
\value{
The YrYr parameter for the given time series
}
\description{
Year on year average
}
\examples{
# Generate an example time series
ts <- seq(-2, 10, by = 0.1) # as a vector of times
ys <- 3 + 2 * ts # plus a vector of values
yryr(ts, ys)
}
\references{
Frazier, R. J., Coops, N. C., Wulder, M. A., Hermosilla, T., & White, J. C. (2018).
Analyzing spatial and temporal variability in short-term rates of post-fire vegetation return
from Landsat time series. Remote Sensing of Environment, 205, 32–45.
\url{https://doi.org/10.1016/j.rse.2017.11.007}
}
