% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{calcNBR}
\alias{calcNBR}
\title{Calculate the Normalized Burn ratio}
\usage{
calcNBR(NIR, SWIR)
}
\arguments{
\item{NIR}{Near Infrored reflectance}

\item{SWIR}{Shortwave Infrared reflectance}
}
\value{
Normalized Burn Ratio
}
\description{
Calculate the Normalized Burn ratio
}
