% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{setPeriod}
\alias{setPeriod}
\title{Fix the time span of a multi-temporal raster stack to a user defined time window}
\usage{
setPeriod(br, startyr, endyr, tempRes, dtsbr)
}
\arguments{
\item{br}{terra raster stack}

\item{tempRes}{temporal resolution of the raster stack}

\item{dtsbr}{vector of dates (Date object) associated with the raster stack}

\item{starttime}{start date of the desired time span (Date object)}

\item{endtime}{end date of the desired time span (Date object)}
}
\value{
terra raster stack with adjusted time span
}
\description{
Fix the time span of a multi-temporal raster stack to a user defined time window
}
